consolidate e2e into 1 package, ensure that do not run tests in parallel (though only having a single package could also cause that)
