+++ Makefile
@@ -13,7 +13,7 @@ test: test-unit test-e2e
 test-unit:
 	go test ./cmd/... ./pkg/...
 test-e2e:
-	KUBERNETES_CONFIG=${KUBECONFIG} go test -timeout 30m -v ./test/e2e/...
+	KUBERNETES_CONFIG=${KUBECONFIG} go test -parallel 1 -timeout 30m -v ./test/e2e/...
 clean:
 	rm -f cluster-samples-operator
